+++ Sources/TorusUtils/Convenience/ED25591.swift
@@ -8,8 +8,8 @@
 import CTweetNacl
 import Foundation
 
-public struct ED25591 {
-    /// Returns ED25591 Keypair
+public struct ED25519 {
+    /// Returns ED25519 Keypair
     /// - Parameter privateKey: Private key returned from `getTorusKey`
     /// - Returns: Returns a tuple containing an ED25519 secretKey (sk) and publicKey (pk)
     public static func getED25519Key(privateKey: String) throws -> (sk: String, pk: String) {
